requirements/base.txt
Django==4.2.*
djangorestframework==3.15.*
drf-spectacular==0.27.*
django-environ==0.11.*
django-cors-headers==4.4.*
psycopg2-binary==2.9.*
celery[redis]==5.3.*
django-celery-beat==2.6.*
firebase-admin==6.5.*
requests==2.31.*
gunicorn==21.*
boto3>=1.34
